+++ src/earthkit/meteo/thermo/array/cape_cin.py
